requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.1
cassandra-driver==3.29.1
httpx>=0.24.0
python-dotenv>=1.0.0
